feat: Implement employee portal with dedicated authentication, navigation, and push notification support.

+++ database/migrations/2026_03_18_000001_create_system_notifications_table.php
@@ -11,11 +11,11 @@ return new class extends Migration
         if (!Schema::hasTable('system_notifications')) {
             Schema::create('system_notifications', function (Blueprint $table) {
                 $table->id();
-                $table->unsignedInteger('sender_id');
+                $table->unsignedBigInteger('sender_id');
                 $table->foreign('sender_id')->references('id')->on('users')->cascadeOnDelete();
                 
                 // null = para todos; si tiene valor, es solo para ese usuario
-                $table->unsignedInteger('target_user_id')->nullable();
+                $table->unsignedBigInteger('target_user_id')->nullable();
                 $table->foreign('target_user_id')->references('id')->on('users')->cascadeOnDelete();
                 
                 $table->string('title');
@@ -30,7 +30,7 @@ return new class extends Migration
             Schema::create('system_notification_reads', function (Blueprint $table) {
                 $table->id();
                 $table->foreignId('notification_id')->constrained('system_notifications')->cascadeOnDelete();
-                $table->unsignedInteger('user_id');
+                $table->unsignedBigInteger('user_id');
                 $table->foreign('user_id')->references('id')->on('users')->cascadeOnDelete();
                 $table->timestamp('read_at');
                 $table->unique(['notification_id', 'user_id']);
